Fix bug in flash()
flash() works by taking an object array that describes color/alpha transitions for a set of objects to be color/alpha tweened, then creating an array that describes the reverse transitions, then repeatedly iterating through the forward and reverse transitions a number of times. I was trying to initialze the reverse transition array by copying the forward transition array, but the assignment operator creates a reference. Had to use some JSON methods to create a copy.